tftordle-be/requirements.txt
Flask
flask-cors
requests
python-dotenv
APScheduler
Pillow
psycopg2-binary